requirements.txt
dimod
dwave-greedy
dwave-hybrid                  
dwave-neal                    
dwave-ocean-sdk               
dwave-system                  
dwave-tabu                    
dwavebinarycsp
matplotlib
numpy
penaltymodel
seaborn
